Added foundations for displaying a unit's Abilities to the user
- Added the AbilityPanel class designed for being displayed in the
InfoPanel
- Added the AbilityPane class for representing a single abiltiy of a
single unit; multiple AbilityPanes are put into a single AbilityPanel
- Added utils package; it currently has two types: IntRef and
IntChangeListener. IntRef stores an int value, and allows code to be run
whenever the value changes. IntChangeListener is a FunctionalInterface
that can be used to supply the code that will be run when an IntRef's
int value changes.
- Modified StepMove to use IntRef instead of int to represent its
distance value.